+++ routes/api/user.php
@@ -2,11 +2,16 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ExpenseController;
+use App\Http\Controllers\ReminderController;
 use Illuminate\Support\Facades\Route;
 
-// expenses
+// expenses routes
 
 Route::get('/expenses', [ExpenseController::class, 'index']);
 Route::post('/expense', [ExpenseController::class, 'store']);
 
-//
+// reminders routes
+
+Route::get('/reminders', [ReminderController::class, 'index']);
+Route::post('/reminder', [ReminderController::class, 'store']);
+Route::post('/reminder/complete', [ReminderController::class, 'markAsCompleted']);