+++ CHANGELOG.md
@@ -8,9 +8,14 @@
  - postman documentation started.
  - New Package google/apiclient installed 
  - Expense module added
+ - Reminder module added
 
 ### Fixed
+ - expense & reminders category using enum classes.
+ - for better development purpose, seeder classes used for seeding the database.
 
 ### Changed
  - enums introduced in App\Enums
+ - enums impleted in db migrations
+
 ### Removed